+++ src/main/java/com/rmr101/campus/repository/CourseRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CourseRepository extends CrudRepository<Course, Integer> {
+public interface CourseRepository extends CrudRepository<Course, Long> {
     List<Course> findByName(String name);
     List<Course> findBySubjectId(int id);
 }
