+++ src/main/java/com/rmr101/campus/repository/SubjectRepository.java
@@ -3,5 +3,5 @@ package com.rmr101.campus.repository;
 import com.rmr101.campus.entity.Subject;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SubjectRepository extends CrudRepository<Subject, Integer> {
+public interface SubjectRepository extends CrudRepository<Subject, Long> {
 }
